requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
xgboost>=1.4.0
torch>=1.9.0
torchvision>=0.10.0
nibabel>=3.2.1
opencv-python>=4.5.0
seaborn>=0.11.0
tqdm>=4.61.0
Pillow>=8.3.0
scikit-image>=0.19.2
scipy>=1.7.0
umap-learn>=0.5.0
